Split out in app purchase parsing test to reduce file size and pacify swiftlint
